requirements.txt
torch
transformers
datasets
numpy
PyYAML
transformer-lens
plotly
wandb
tiktoken
tqdm
scikit-learn
accelerate
matplotlib
